[REFACTOR]Framework: Creation of VersionHelpers To be used when wanting the version of the application or Extensions
